refactor: features/nodeManagement 제거
