+++ frontend/techpick/src/components/nodeManagement/api/folder/useCreateFolder.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { ApiFolderData } from '@/types/ApiTypes';
-import { postFolder } from '@/features/nodeManagement/api/folder/folderQueryFunctions';
+import { postFolder } from '@/components/nodeManagement/api/folder/folderQueryFunctions';
 
 export const useCreateFolder = () => {
   return useMutation<ApiFolderData, Error, string>({
