+++ frontend/techpick/src/components/nodeManagement/api/folder/useGetDefaultFolderData.ts
@@ -1,6 +1,6 @@
 import { ApiDefaultFolderIdData } from '@/types/ApiTypes';
 import { useQuery } from '@tanstack/react-query';
-import { getFoldersIdData } from '@/features/nodeManagement/api/folder/folderQueryFunctions';
+import { getFoldersIdData } from '@/components/nodeManagement/api/folder/folderQueryFunctions';
 
 export const useGetDefaultFolderData = () => {
   return useQuery<ApiDefaultFolderIdData>({
