+++ frontend/techpick/src/components/nodeManagement/api/folder/useGetRootAndRecycleBinData.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getRootAndRecycleBinData } from '@/features/nodeManagement/api/folder/folderQueryFunctions';
+import { getRootAndRecycleBinData } from '@/components/nodeManagement/api/folder/folderQueryFunctions';
 
 export const useGetRootAndRecycleBinData = () => {
   return useQuery({
