+++ frontend/techpick/src/components/nodeManagement/api/folder/useMoveFolder.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { putFolderMove } from '@/features/nodeManagement/api/folder/folderQueryFunctions';
+import { putFolderMove } from '@/components/nodeManagement/api/folder/folderQueryFunctions';
 
 export const useMoveFolder = () => {
   return useMutation({
