+++ frontend/techpick/src/components/nodeManagement/api/folder/useRenameFolder.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { putFolderRename } from '@/features/nodeManagement/api/folder/folderQueryFunctions';
+import { putFolderRename } from '@/components/nodeManagement/api/folder/folderQueryFunctions';
 
 export const useRenameFolder = () => {
   return useMutation({
