+++ frontend/techpick/src/components/nodeManagement/api/pick/pickQueryFunctions.ts
