+++ frontend/techpick/src/components/nodeManagement/api/pick/useCreatePick.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { ApiPickData, ApiPickRequestType } from '@/types/ApiTypes';
-import { postPick } from '@/features/nodeManagement/api/pick/pickQueryFunctions';
+import { postPick } from '@/components/nodeManagement/api/pick/pickQueryFunctions';
 
 export const useCreatePick = () => {
   return useMutation<ApiPickData, Error, ApiPickRequestType>({
