+++ frontend/techpick/src/components/nodeManagement/api/pick/useDeletePick.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { deletePick } from '@/features/nodeManagement/api/pick/pickQueryFunctions';
+import { deletePick } from '@/components/nodeManagement/api/pick/pickQueryFunctions';
 
 export const useDeletePick = () => {
   return useMutation({
