+++ frontend/techpick/src/components/nodeManagement/api/pick/useMovePick.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { putPickMove } from '@/features/nodeManagement/api/pick/pickQueryFunctions';
+import { putPickMove } from '@/components/nodeManagement/api/pick/pickQueryFunctions';
 
 export const useMovePick = () => {
   return useMutation({
