+++ frontend/techpick/src/components/nodeManagement/ui/Pick.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { NodeApi } from 'react-arborist';
 
 import Image from 'next/image';
-import { linkWrapper } from '@/features/nodeManagement/ui/pick.css';
-import { useDragHook } from '@/features/nodeManagement/hooks/useDragHook';
+import { linkWrapper } from '@/components/nodeManagement/ui/pick.css';
+import { useDragHook } from '@/components/nodeManagement/hooks/useDragHook';
 
 // 여기에 PickCard카드될 예정입니다.
 export function Pick({ node }: { node: NodeApi }) {
