+++ frontend/techpick/src/components/nodeManagement/ui/pick.css.ts
