+++ frontend/techpick/src/components/nodeManagement/utils/addNodeToStructure.ts
