+++ frontend/techpick/src/components/nodeManagement/utils/addRecycleBinToStructure.ts
@@ -1,6 +1,6 @@
 import { NodeData } from '@/types';
-import { addNodeToStructure } from '@/features/nodeManagement/utils/addNodeToStructure';
-import { getNewIdFromStructure } from '@/features/nodeManagement/utils/getNewIdFromStructure';
+import { addNodeToStructure } from '@/components/nodeManagement/utils/addNodeToStructure';
+import { getNewIdFromStructure } from '@/components/nodeManagement/utils/getNewIdFromStructure';
 import { ApiStructureData } from '@/types/ApiTypes';
 
 export function addRecycleBinToStructure(
