+++ frontend/techpick/src/components/nodeManagement/utils/getCurrentTreeTypeByNode.ts
