+++ frontend/techpick/src/components/nodeManagement/utils/getNewIdFromStructure.ts
