+++ frontend/techpick/src/components/nodeManagement/utils/moveNode.ts
