+++ frontend/techpick/src/components/nodeManagement/utils/removeByIdFromStructure.ts
