+++ frontend/techpick/src/widgets/ContextMenu/EditorContextMenu.tsx
@@ -11,8 +11,8 @@ import {
   ContextMenuSubTrigger,
 } from './ContextMenu.css';
 import { useTreeStore } from '@/stores/treeStore';
-import { getCurrentTreeTypeByNode } from '@/features/nodeManagement/utils/getCurrentTreeTypeByNode';
-import { useRestoreNode } from '@/features/nodeManagement/hooks/useRestoreNode';
+import { getCurrentTreeTypeByNode } from '@/components/nodeManagement/utils/getCurrentTreeTypeByNode';
+import { useRestoreNode } from '@/components/nodeManagement/hooks/useRestoreNode';
 import toast from 'react-hot-toast';
 
 interface ContextMenuWrapperProps {
